Changed dotenv integration, refactored bootstrap
