+++ .env.example
@@ -1,11 +1,19 @@
 PORT=3000
+
 LOG_DIR=./logs
+
+# Database 
+DB_TYPE=
 DB_HOST=localhost
-DB_PORT=27017
+DB_PORT=
+DB_USERNAME=
+DB_PASSWORD=
 DB_NAME=fliptask
+
+# Auth
 JWT_KEY=gdsfhjsbjhbdjhsdji68743687y497
-API_URL=localhost:3000
-HOST_URL=http://localhost:3000
+
+# Email
 MAILER_USERNAME=
 MAILER_PASSWORD=
 MAILER_SERVICE_NAME=
