+++ README.md
@@ -15,18 +15,18 @@ Steps to follow
 1.  Clone Fliptask-API
 2.  Change the current directory to `fliptask-api`.
 ```
-cd tasify
+cd fliptask-api
 ```
-3. Run the install script `install.sh`
+3. Copy `.env.example` as `.env`, make required changes.
+
+4. Run the install script `install.sh`
 ```
 bash ./install.sh
 ```
-4. To start the server run:
+5. To start the server run:
 ```
 npm run dev 
 ```
-
-> open <http://localhost:3000> in browser.
 
 ### Possible Errors and its solutions
 If get any `error` in running `install.sh`,  example :
