+++ server/bootstrap/Controller.js
@@ -9,7 +9,7 @@ const init = async () => {
             const dep = file.split(".")[0];
             acc[dep] = res;
             global[dep] = acc[dep];
-            console.info(`[INFO] Dependency [${file}] loaded from [controllers]`);
+            // console.info(`[INFO] Dependency [${file}] loaded from [controllers]`);
             return acc;
         } catch (e) {
             console.error(`[Error] Unable to load ${file}`, e);
