+++ server/bootstrap/Middleware.js
@@ -9,7 +9,7 @@ const init = async () => {
             const dep = file.split(".")[0];
             acc[dep] = res;
             global[dep] = acc[dep];
-            console.info(`[INFO] Dependency [${file}] loaded from [middlewares]`);
+            // console.info(`[INFO] Dependency [${file}] loaded from [middlewares]`);
             return acc;
         } catch (e) {
             console.error(`[Error] Unable to load ${file}`, e);
