+++ server/index.js
@@ -1,3 +1,49 @@
-const { loadApp } = require("./bootstrap");
+require("dotenv").config();
+
+const loadDependency = async (type, dep, moduleObj) => {
+    try {
+        const res = await require(`./${type}/${dep}`);
+        global[dep] = typeof (res) === "function" ? await res(moduleObj) : await res;
+        moduleObj[dep] = global[dep];
+        // console.info(`[INFO] Dependency [${dep}] loaded from [${type}]`);
+        return global[dep];
+    } catch (e) {
+        console.error(`[Error] Unable to load ${dep}\n`, e);
+        return e;
+    }
+};
+const configuration = {
+    services: [
+        "MailerService"
+    ],
+    bootstrap: [
+        "Multer",
+        "Mailer",
+        "Logger",
+        "DB",
+        "Model",
+        "Middleware",
+        "Controller",
+        "Route",
+        "Express"
+    ]
+};
+const loadConfig = async () => {
+    const moduleObj = {};
+    const typeList = Object.keys(configuration);
+    for (let i = 0; i < typeList.length; i++) {
+        const type = typeList[i];
+        // console.log(`[#################################******** Loading ${type} *********#################################]`);
+        for (let j = 0; j < configuration[type].length; j++) {
+            const name = configuration[type][j];
+            // eslint-disable-next-line no-await-in-loop
+            await loadDependency(type, name, moduleObj);
+        }
+    }
+};
+
+const loadApp = async () => {
+    await loadConfig();
+};
 
 loadApp();
